rename RHFInput to Controller
